+++ src/main/java/com/teamtea/eclipticseasons/api/constant/climate/ISnowTerm.java
@@ -8,7 +8,16 @@ public interface ISnowTerm {
 
     SolarTerm getEnd();
 
+    default ISnowTerm cast(float tempChange) {
+        return this;
+    }
+
     default boolean maySnow(SolarTerm solarTerm) {
         return solarTerm.isInTerms(getStart(), getEnd());
+    }
+
+    @Deprecated
+    default boolean maySnow(SolarTerm solarTerm, float tempChange) {
+        return cast(tempChange).maySnow(solarTerm);
     }
 }