+++ src/main/java/com/teamtea/eclipticseasons/common/network/message/BroomUseMessage.java
@@ -4,7 +4,7 @@ package com.teamtea.eclipticseasons.common.network.message;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.FriendlyByteBuf;
 
-public final class BroomUseMessage {
+public class BroomUseMessage {
     public final BlockPos blockPos;
     public final long time;